POST messages et mapping. Amelioration type retour

+++ demo/src/main/java/com/openclassrooms/backend/dto/MessageRequestDTO.java
@@ -3,8 +3,8 @@ package com.openclassrooms.backend.dto;
 public class MessageRequestDTO {
 
   private String message;
-  private int user_id;
-  private int rental_id;
+  private Long user_id;
+  private Long rental_id;
 
 
   public String getMessage() {
@@ -15,19 +15,19 @@ public class MessageRequestDTO {
     this.message = message;
   }
 
-  public int getUser_id() {
+  public Long getUser_id() {
     return user_id;
   }
 
-  public void setUser_id(int user_id) {
+  public void setUser_id(Long user_id) {
     this.user_id = user_id;
   }
 
-  public int getRental_id() {
+  public Long getRental_id() {
     return rental_id;
   }
 
-  public void setRental_id(int rental_id) {
+  public void setRental_id(Long rental_id) {
     this.rental_id = rental_id;
   }
 }
